requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-socketio==5.3.6
eventlet==0.35.1
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
dnspython==2.7.0
cryptography==43.0.0
paramiko==3.4.0
pytest==8.3.2
